+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Query to unassign a group from a user
 - Query to unassign a role from a group
 - Query to unassign a marking definition from a group
+- Query to unassign a capability from a role
 
 ## [0.2.0] - 2024-12-18